Merged PR 39187: Add endpoint for product brand
## What's being changed

We've added a new endpoint to retrieve a product's nominated brand attribute value.

## Why it's being changed

This needs to be a new endpoint on our side, since merchants can configure different brand attributes (e.g. Manufacturer, Brand, My brand etc.) in our module config.

## How to review / test this change

- Set an attribute to be the brand value for products in a given scope (Dotdigital > Configuration)
- Ensure this attribute is part of the attribute set of the product you're testing
- Run setup:upgrade
- In the GraphQl playground GUI, use this query
```
    query getProductBrand($product_id: String!) {
        productBrand(product_id: $product_id) {
            value
        }
    }
```
- Use query params as per screenshot
![image (2).png](https://dev.azure.com/dotdigital/b2ba4bc1-4a9a-48f7-a69a-5cf84134b7c4/_apis/git/repositories/88ee8b5e-f928-4591-bc1f-8106b7c53d04/pullRequests/39187/attachments/image%20%282%29.png)
- Run the query and receive a value in response

## Notes

'Manufacturer' is a dropdown attribute in Venia sample data, therefore the value returned will actually be the id of the option selected. We don't currently support this in Magento (the brand value needs to be a text field).

Related work items: #182778